core.fs: Fix casts for non-64-bit architectures

+++ core/core.fs/main.c
@@ -18,44 +18,44 @@ uint64_t send(uint16_t _cmd, void* data) {
 	// access commands
 
 	if (cmd == CMD_OPEN) {
-		const char* drive = (const void*) args[0];
-		const char* path  = (const void*) args[1];
+		const char* drive = (const void*) (intptr_t) args[0];
+		const char* path  = (const void*) (intptr_t) args[1];
 
 		flags_t flags = args[2];
 
-		return (uint64_t) fs_open(drive, path, flags);
+		return (uint64_t) (intptr_t) fs_open(drive, path, flags);
 	}
 
 	else if (cmd == CMD_CLOSE) {
-		descr_t* descr = (void*) args[0];
+		descr_t* descr = (void*) (intptr_t) args[0];
 		return fs_close(descr);
 	}
 
 	// info commands
 
 	else if (cmd == CMD_SIZE) {
-		descr_t* descr = (void*) args[0];
-		return (uint64_t) fs_size(descr);
+		descr_t* descr = (void*) (intptr_t) args[0];
+		return (uint64_t) (intptr_t) fs_size(descr);
 	}
 
 	else if (cmd == CMD_MMAP) {
-		descr_t* descr = (void*) args[0];
-		return (uint64_t) fs_mmap(descr);
+		descr_t* descr = (void*) (intptr_t) args[0];
+		return (uint64_t) (intptr_t) fs_mmap(descr);
 	}
 
 	// stream commands
 
 	else if (cmd == CMD_READ) {
-		descr_t* descr = (void*) args[0];
-		void* buf = (void*) args[1];
+		descr_t* descr = (void*) (intptr_t) args[0];
+		void* buf = (void*) (intptr_t) args[1];
 		size_t len = args[2];
 
 		return fs_read(descr, buf, len);
 	}
 
 	else if (cmd == CMD_WRITE) {
-		descr_t* descr = (void*) args[0];
-		const void* buf = (const void*) args[1];
+		descr_t* descr = (void*) (intptr_t) args[0];
+		const void* buf = (const void*) (intptr_t) args[1];
 		size_t len = args[2];
 
 		return fs_write(descr, buf, len);
